requirements.txt
fastapi[all]
websockets
aiohttp
python-dotenv
ujson
aiobotocore
loguru
jinja2
motor
python-multipart